Move deprecation warnings for old state fields so they don't always trigger.
