+++ code/lib/manager-api/src/index.tsx
@@ -38,6 +38,7 @@ import {
 } from '@storybook/core-events';
 import type { RouterData } from '@storybook/router';
 import type { Listener } from '@storybook/channels';
+import { deprecate } from '@storybook/client-logger';
 
 import { createContext } from './context';
 import type { Options } from './store';
@@ -318,7 +319,23 @@ function ManagerConsumer<P = Combo>({
 
 export function useStorybookState(): State {
   const { state } = useContext(ManagerContext);
-  return state;
+  return {
+    ...state,
+
+    // deprecated fields for back-compat
+    get storiesHash() {
+      deprecate('state.storiesHash is deprecated, please use state.index');
+      return this.index || {};
+    },
+    get storiesConfigured() {
+      deprecate('state.storiesConfigured is deprecated, please use state.previewInitialized');
+      return this.previewInitialized;
+    },
+    get storiesFailed() {
+      deprecate('state.storiesFailed is deprecated, please use state.indexError');
+      return this.indexError;
+    },
+  };
 }
 export function useStorybookApi(): API {
   const { api } = useContext(ManagerContext);
